Update v1
Wrote a code that paints the French flag.
The code can be written with fewer lines if you use a for loop inside a for loop in which the first for loop iterates over the colors and he second for loop iterates over the width and height of the flag, respectively.